feat: support enum type and multiple namespaces

+++ common/orka-utils.h
@@ -50,7 +50,8 @@ void orka_sleep_ms(const int64_t delay_ms);
 uint64_t orka_timestamp_ms();
 void orka_timestamp_str(char *p_str, int len);
 
-int orka_str_to_ntl(
+size_t
+orka_str_to_ntl(
   char *str,
   size_t size,
   struct ntl_deserializer * deserializer);
